refactor: Rename app package to calendar and update references
